refactored temp file
