+++ src/NuGet.Core/NuGet.Common/AsyncLazy.cs
